scrapers/requirements.txt
beautifulsoup4==4.11.1
black
brotli
feedparser
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
icalendar
ics
isort
lxml
nest-asyncio
prisma
python-dateutil==2.8.2
python-dotenv
pytz==2022.6
requests==2.28.1
selenium
webdriver-manager
